Rename `formatClassName.ts` to `format.ts` to unify utils export
