+++ src/components/common/Saturation.tsx
@@ -5,7 +5,7 @@ import Interactive from "./Interactive";
 import styles from "../../css/styles.css";
 import { HSV } from "../../types";
 import { hsvToHslString } from "../../utils/convert";
-import formatClassName from "../../utils/formatClassName";
+import { formatClassName } from "../../utils/format";
 
 interface Props {
   hsv: HSV;
