+++ src/utils/format.ts
@@ -1,0 +1,1 @@
+export const formatClassName = (names: string[]): string => names.filter(Boolean).join(" ");
