--- src/utils/formatClassName.ts
@@ -1,3 +1,0 @@
-const formatClassName = (names: string[]): string => names.filter(Boolean).join(" ");
-
-export default formatClassName;
